Completed MVP
MVP of Simple Calculator. Takes an input and an operator as many times as needed then converts that into a string that evals into an equation.

The result of that equation can even be used to form the next equation.

TODO
fix bug where pushing any other number just adds to the result of the last equation instead of clearing the display and entering in that value.