Refactor to remove redundant variable
having Operators stored in a separate array to "zip" together with the numbers at the point of calculation is redundant.

The Operator array was first created because I thought I could pass in operators as they were instead of as strings. Given the numbers are also passed in as strings, why not just put it all in the one array, making a simpler loop without needing to manage two different array lengths.

The only consideration for this new approach is the order in which you add to the array. It should be: number, then operator. E.G "1, +"