Change from Array to Map
Changed from storing data in an array to a Map. This was the original plan, I was just not certain how to go about it originally. Now, after some experimenting, I know how to dynamically generate a key and store both the text and the state of the checkbox.

Note: I'll need to set up a function to slice the two when I retrieve them.