Added localStorage
Scrapped all of the map stuff and went back to an array. I wanted to play with maps, I did that and found it wasn't fit for this purpose.

I can parse the array and load tasks, in order, and with their checkbox ticked or not from localStorage.

But for some reason the array gets bigger with a duplicate value when it's loaded. I'll need to trace where that's happening.